[CAS-1304] Adding file attachments
