+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/attachment/AttachmentViewHolderFactory.kt
@@ -1,0 +1,4 @@
+package io.getstream.chat.android.ui.message.list.adapter.viewholder.attachment
+
+public open class AttachmentViewHolderFactory {
+}
